Extract Scala2 macro code as Impl traits
